packages/pipelines/pyproject.toml
[project]
name = "rhoai-mcp-pipelines"
version = "0.1.0"
description = "Pipelines component for RHOAI MCP - Data Science Pipelines management"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Adam Miller", email = "admiller@redhat.com" }
]
maintainers = [
    { name = "Pipelines Team", email = "pipelines-team@redhat.com" }
]
requires-python = ">=3.10"
keywords = ["mcp", "openshift", "rhoai", "pipelines", "kubeflow"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "rhoai-mcp-core>=0.1.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.entry-points."rhoai_mcp.plugins"]
pipelines = "rhoai_mcp_pipelines.plugin:create_plugin"

[project.urls]
Homepage = "https://github.com/admiller/rhoai-mcp-prototype"
Repository = "https://github.com/admiller/rhoai-mcp-prototype"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rhoai_mcp_pipelines"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint.isort]
known-first-party = ["rhoai_mcp_pipelines", "rhoai_mcp_core"]
